esquema básico de autenticação

+++ cashcontroller-api/src/main/java/br/com/cashcontroller/dto/OperacaoRendaVariavelDTO.java
@@ -18,7 +18,7 @@ public class OperacaoRendaVariavelDTO {
 	private AtivoDTO ativoDto;
 	private double valorCorretagem;
 	private LocalDate dataOperacao;
-	private Long quantidadeNegociada;
+	private int quantidadeNegociada;
 	private TipoOperacaoDTO tipoOperacaoDto;
 	private double custoTotal;
 	private double valorTotal;
